commit#12: created an image noising and then image denoising project. the 32x32 resolution images were loaded from the CIFAR 10 dataset. then we applied noisiness on the image.
Once that was done we built a Neural Network with tensorflow and keras with 2 layers of hidden layers and one output layer. with that we denoised the noised-image. Once all this was done we trained the model with 10 epochs. and gave it our validation data.
Finally we plotted the all copies of our a.original image b. noisy image c. de-noised image using matplotlib . Total Time taken = 4 hours